make dtype explcit to fix appveyor test error

+++ pandas/tests/scalar/test_timedelta.py
@@ -191,7 +191,7 @@ class TestTimedeltaArithmetic(object):
         assert td // np.array(scalar.to_timedelta64()) == 1
 
         res = (3 * td) // np.array([scalar.to_timedelta64()])
-        expected = np.array([3])
+        expected = np.array([3], dtype=np.int64)
         tm.assert_numpy_array_equal(res, expected)
 
         res = (10 * td) // np.array([scalar.to_timedelta64(),
